refactor: replace read packet permission utility function and update related logic
